fixup! riscv: Emulate one-byte and two-byte cmpxchg

+++ arch/riscv/include/asm/cmpxchg.h
@@ -172,11 +172,11 @@
 	register unsigned int __rc;					\
 	switch (size) {							\
 	case 1:								\
-		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;							\
 	case 2:								\
 		break;							\
-		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 	case 4:								\
 		__asm__ __volatile__ (					\
 			"0:	lr.w %0, %2\n"				\
